splash

+++ src/app/co-users/co-users/co-users.component.ts
@@ -1,8 +1,10 @@
 import { Component, OnInit, HostBinding } from '@angular/core';
-import { CoUser } from '../../models/coUser';
+import { CoUser } from '../coUser';
 import { ApiService } from '../../services/api.service';
 import { routerAnimation } from '../../animations';
 import { GobackService } from '../../services/goback.service';
+import { DialogService } from '../../services/dialog.service';
+import { CoUserApiService } from '../co-user-api.service';
 
 @Component({
   selector: 'app-co-users',
@@ -18,7 +20,7 @@ export class CoUsersComponent implements OnInit {
 
   coUsers: CoUser[] = [];
 
-  constructor(private api: ApiService, private goback:GobackService) { }
+  constructor(private api: CoUserApiService, private goback:GobackService, private dialog: DialogService) { }
 
   ngOnInit() {
     this.goback.urlInit();
@@ -40,11 +42,19 @@ export class CoUsersComponent implements OnInit {
   }
 
   delete(coUser: CoUser) {
-    this.api.deleteCoUser(coUser).subscribe(
-      data => {
-        if (data.success) {
-          this.coUsers = this.coUsers.filter(h => h.id !== coUser.id);
+    this.dialog.confirm("Are you sure want to delete this Co-User?").subscribe(
+      confirm => {
+        if (!confirm) {
+          return;
         }
+
+        this.api.deleteCoUser(coUser).subscribe(
+          data => {
+            if (data.success) {
+              this.coUsers = this.coUsers.filter(h => h.id !== coUser.id);
+            }
+          }
+        );
       }
     );
   }
